feat(ZEMA-1764): (refactor module for cardholder)
