+++ lib/m2y_fast/xml/xml_response_parser.rb
@@ -3,38 +3,6 @@
 module M2yFast
   class XmlResponseParser
 
-    ### CARDHOLDER ###
-
-    def self.get_cardholder_response(json)
-      begin
-        xml_str = json[:consulta_portador_response][:return]
-        codigo_retorno = xml_str.split("<codigo_retorno>").last.split("</codigo_retorno>").first.to_i
-        { error: codigo_retorno != 0, code: codigo_retorno }
-      rescue
-        { error: true }
-      end
-    end
-
-    def self.get_user_registration_response(json)
-      begin
-        xml_str = json[:dados_gral_portador_response][:return]
-        codigo_retorno = xml_str.split("<codigo_retorno>").last.split("</codigo_retorno>").first.to_i
-        { error: codigo_retorno != 0, code: codigo_retorno }
-      rescue
-        { error: true }
-      end
-    end
-
-    def self.update_user_registration_response(json)
-      begin
-        xml_str = json[:alteracao_portador2_response][:return]
-        codigo_retorno = xml_str.split("<codigo_retorno>").last.split("</codigo_retorno>").first.to_i
-        { error: (codigo_retorno != 0 && codigo_retorno != 1000), code: codigo_retorno }
-      rescue
-        { error: true }
-      end
-    end
-
     ### CARD ###
 
     def self.request_card_response(json, body)
